Refactor: Change workflow_lead_id to unsignedBigInteger

+++ database/migrations/2025_10_07_000001_add_workflow_lead_id_to_activities_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('activities', function (Blueprint $table) {
-            $table->unsignedInteger('workflow_lead_id')->nullable()->after('lead_id');
+            $table->unsignedBigInteger('workflow_lead_id')->nullable()->after('lead_id');
             $table->foreign('workflow_lead_id')->references('id')->on('salesleads')->onDelete('cascade');
         });
     }
